backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
PyMySQL==1.1.1
cryptography==45.0.6
python-multipart==0.0.20
razorpay==1.4.2
PyJWT==2.10.1
setuptools<81
